// types/booking.ts
export interface BookingDetails {
  reservationDetailID?: number;
  reservationMasterID?: number;
  reservationID?: number;
  nameID?: number;
  guestName?: string;
  guest?: string;
  roomNumber?: string;
  currencyCode?: string;
  roomID?: number;
  mealPlan?: string;
  checkIn?: string;
  checkOut?: string;
  nights?: number;
  status?: string;
}